Fixup: correct the shared key generation as per SEC 1 section 3.3.1
It appears that the go-ethereum and its derivatives only ever tested the P-256 and S-256 curves for ECIES.
The tests for other curves are missing (apart from the params check).
What's worse, the P-521 curve shared key generation panics with the current parameters.

The SEC 1 section 3.3.1 says that the ECDH shared key should be taken as `scalar(point(pubkey), privkey).X`.
The Golang's standard library does the same: https://github.com/golang/go/blob/go1.19.5/src/crypto/tls/key_schedule.go#L167
Furthermore, the HSMs do the same (at least for the P-256 curve, as verified in LmP).

Thus, it looks strange at first place why the go-ethereum decided to pick the last N bytes of the shared key.
Especially, given that the `skLen, macLen` parameters to `GenerateShared` function were specially crafted for
the P-256 curve in such a way that the full shared key is returned (equal to what SEC 1 says).
This even worked for the P-384 curve in go-ethereum (but not a fork, see https://github.com/foundriesio/go-ecies/pull/2).

But the code panics for the P-521 curve: it's ECDH shared key is 66 bytes long while the code tried to get 64 bytes.
Ironically, due to error in bytes calculus (originated in go-ethereum) code tries to fetch byte at index -2 and panics.

This change brings the `GenerateShared` function to conform to the SEC 1 section 3.3.1 (ECDH shared key).
For that, the `skLen, macLen` parameters are not needed... they are actually needed later to be passed into the KDF.
So, the function signature if changed to remove unused arguments (that will also make the PKCS11 integration easier).

The changed `GenerateShared` function still passes the test vectors in https://github.com/foundriesio/go-ecies/pull/4
which were generated using the previous implementation (including the P-521 curve).
The reason why they pass for P-521 is because the GenerateShared tests ignore curve parameters and use a max key length.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>