Fixup: accurately calculate the marshalled public key length
I have no idea where did the go-ethereum take that formula `(bitsize+7)/4`.
Maybe, they assumed that BitSize is always a multiple of 8, thus it equals `2*((bitsize+7)/8)`.
But that assumption is wrong even for the go-ethereum codebase as they support the P-521 curve.

So, when I enabled encrypt/decrypt tests to for all curves (not only the default P-256)...
the decryption started to fail with "invalid public key" error for the P-521 curve.

Fortunately, the correct formula is well-known from the SEC 1 section 2.3.4,
and can be copied from Golang. It gives a difference 133 vs 132 bytes for the P-521 curve.

In scope of this library we're only interested in the non-compressed octet stream (case 4);
but I copied the formula for the compressed case as well (which was even more wrong).

Vector tests for encryption & decryption added in https://github.com/foundriesio/go-ecies/pull/4.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>